requirements.txt
streamlit
mysql-connector-python
pandas
faker
plotly
